+++ lib/eve_online/esi/models/dogma_effects.rb
@@ -7,6 +7,8 @@ module EveOnline
         def dogma_effects
           output = []
 
+          return output if !options.is_a?(Array)
+
           options.each do |dogma_effect_short|
             output << DogmaEffectShort.new(dogma_effect_short)
           end