+++ spec/eve_online/esi/models/dogma_effects_spec.rb
@@ -14,14 +14,22 @@ describe EveOnline::ESI::Models::DogmaEffects do
   end
 
   describe '#dogma_effects' do
-    let(:dogma_effect_short) { double }
+    context 'when options is array' do
+      let(:dogma_effect_short) { double }
 
-    let(:option) { double }
+      let(:option) { double }
 
-    let(:options) { [option] }
+      let(:options) { [option] }
 
-    before { expect(DogmaEffectShort).to receive(:new).with(option).and_return(dogma_effect_short) }
+      before { expect(EveOnline::ESI::Models::DogmaEffectShort).to receive(:new).with(option).and_return(dogma_effect_short) }
 
-    before { expect(subject.dogma_effects).to eq([dogma_effect_short]) }
+      before { expect(subject.dogma_effects).to eq([dogma_effect_short]) }
+    end
+
+    context 'when options is nil' do
+      let(:options) { nil }
+
+      before { expect(subject.dogma_effects).to eq([]) }
+    end
   end
 end